refactor: extract shared CommonSearchViewProps for Search views
