+++ src/components/Search/TaskSearchView.tsx
@@ -1,18 +1,13 @@
 import type {ExtendedTargetedEvent} from '@components/SelectionList/ListItem/types';
 
-import type {TransactionPreviewData} from '@libs/actions/Search';
-import type {ModifiedMouseEvent} from '@libs/Navigation/helpers/openInternalRouteInNewTab';
-
 import CONST from '@src/CONST';
-import type {Transaction} from '@src/types/onyx';
 
-import type {ForwardedRef} from 'react';
-import type {NativeScrollEvent, NativeSyntheticEvent, StyleProp, ViewStyle} from 'react-native';
+import type {NativeSyntheticEvent} from 'react-native';
 
 import React, {useImperativeHandle} from 'react';
 
 import type {SearchListItem} from './SearchList/ListItem/types';
-import type {SearchColumnType, SearchQueryJSON} from './types';
+import type {CommonSearchViewProps} from './searchViewProps';
 
 import useSearchListViewState from './hooks/useSearchListViewState';
 import AnimatedExitRow from './primitives/AnimatedExitRow';
@@ -21,66 +16,7 @@ import BaseSearchList from './SearchList/BaseSearchList';
 import TaskListItem from './SearchList/ListItem/TaskListItem';
 import SearchListViewLayout from './SearchListViewLayout';
 
-/** Imperative handle the router uses for highlight-driven scrolling (mirrors SearchList's handle). */
-type SearchListHandle = {
-    scrollToIndex: (index: number, animated?: boolean) => void;
-};
-
-type TaskSearchViewProps = {
-    /** The task search query. */
-    queryJSON: SearchQueryJSON;
-
-    /** The sorted rows to render (from the router's useSearchSnapshot). */
-    data: SearchListItem[];
-
-    /** The columns to render in the list (drives the table min-width and header). */
-    columns: SearchColumnType[];
-
-    /** Whether the list supports multi-select. */
-    canSelectMultiple: boolean;
-
-    /** Whether the action column uses its wider variant. */
-    isActionColumnWide: boolean;
-
-    /** Whether mobile selection mode is on. */
-    isMobileSelectionModeEnabled: boolean;
-
-    /** The column header element (undefined on narrow layouts). */
-    SearchTableHeader?: React.JSX.Element;
-
-    /** Whether a table header bar is shown above the list. */
-    tableHeaderVisible: boolean;
-
-    /** Whether everything has been loaded (gates the fully-checked select-all state). */
-    hasLoadedAllTransactions?: boolean;
-
-    /** Rows flagged for the post-create highlight animation (feeds BaseSearchList extraData). */
-    newTransactions: Transaction[];
-
-    /** The navigation handler for a row tap (owned by the router). */
-    onSelectRow: (item: SearchListItem, transactionPreviewData?: TransactionPreviewData, event?: ModifiedMouseEvent) => void;
-
-    /** The list footer (pagination / pending skeleton). */
-    ListFooterComponent?: React.JSX.Element;
-
-    /** Fires when the list scrolls near its end (router's fetchMoreResults). */
-    onEndReached: () => void;
-
-    /** Fires on the list's first layout and on layout changes. */
-    onLayout: () => void;
-
-    /** Scroll handler forwarded to the list. */
-    onScroll?: (event: NativeSyntheticEvent<NativeScrollEvent>) => void;
-
-    /** Content container style for the list. */
-    contentContainerStyle: StyleProp<ViewStyle>;
-
-    /** Outer container style for the list wrapper. */
-    containerStyle: StyleProp<ViewStyle>;
-
-    /** Imperative handle for highlight-driven scrolling, set by the router. */
-    ref?: ForwardedRef<SearchListHandle>;
-};
+type TaskSearchViewProps = CommonSearchViewProps;
 
 const keyExtractor = (item: SearchListItem, index: number) => item.keyForList ?? `${index}`;
